local-rag-pdf-qa/requirements.txt
# Dependencies for the project
streamlit==1.28.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyMuPDF==1.23.5
pytesseract==0.3.10
pdf2image==1.16.3
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.10.1
isort==5.12.0
flake8==6.1.0